+++ cas-server-core-configuration/src/main/java/org/apereo/cas/configuration/model/core/authentication/PrincipalAttributesProperties.java
@@ -86,6 +86,32 @@ public class PrincipalAttributesProperties {
     }
 
     public static class Ldap extends AbstractLdapProperties {
+        private boolean subtreeSearch = true;
+        private String baseDn;
+        private String userFilter;
 
+        public String getBaseDn() {
+            return baseDn;
+        }
+
+        public void setBaseDn(final String baseDn) {
+            this.baseDn = baseDn;
+        }
+
+        public String getUserFilter() {
+            return userFilter;
+        }
+
+        public void setUserFilter(final String userFilter) {
+            this.userFilter = userFilter;
+        }
+
+        public boolean isSubtreeSearch() {
+            return subtreeSearch;
+        }
+
+        public void setSubtreeSearch(final boolean subtreeSearch) {
+            this.subtreeSearch = subtreeSearch;
+        }
     }
 }
