+++ cas-server-support-ldap-service-registry/src/main/java/org/apereo/cas/adaptors/ldap/services/config/LdapServiceRegistryConfiguration.java
@@ -4,6 +4,7 @@ import org.apereo.cas.adaptors.ldap.services.DefaultLdapRegisteredServiceMapper;
 import org.apereo.cas.adaptors.ldap.services.LdapRegisteredServiceMapper;
 import org.apereo.cas.adaptors.ldap.services.LdapServiceRegistryDao;
 import org.apereo.cas.configuration.CasConfigurationProperties;
+import org.apereo.cas.configuration.support.Beans;
 import org.apereo.cas.services.ServiceRegistryDao;
 import org.ldaptive.ConnectionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,13 +28,7 @@ public class LdapServiceRegistryConfiguration {
 
     @Autowired
     private CasConfigurationProperties casProperties;
-
-
-    @Nullable
-    @Autowired(required = false)
-    @Qualifier("ldapServiceRegistryConnectionFactory")
-    private ConnectionFactory connectionFactory;
-
+    
     @Bean
     @RefreshScope
     public LdapRegisteredServiceMapper ldapServiceRegistryMapper() {
@@ -45,6 +40,10 @@ public class LdapServiceRegistryConfiguration {
     public ServiceRegistryDao ldapServiceRegistryDao() {
         final LdapServiceRegistryDao r = new LdapServiceRegistryDao();
 
+        final ConnectionFactory connectionFactory = Beans.newPooledConnectionFactory(
+                casProperties.getServiceRegistry().getLdap()
+        );
+        
         r.setConnectionFactory(connectionFactory);
         r.setLdapServiceMapper(ldapServiceRegistryMapper());
         r.setBaseDn(casProperties.getServiceRegistry().getLdap().getBaseDn());
